rebased patch  on latest master
